Manual Query https://testing-library.com/docs/queries/about/#manual-queries
On top of the queries provided by the testing library,
you can use the regular querySelector DOM API to query elements.
Note that using this as an escape hatch to query by class or id
is not recommended because they are invisible to the user.
Use a testid if you have to, to make your intention to fall back
to non-semantic queries clear and establish a stable API contract
in the HTML.

// @testing-library/react
const {container} = render(<MyComponent />)
const foo = container.querySelector('[data-foo="bar"]')